--- Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/01 - teams-bot2/teams-bot2/App_Start/WebApiConfig.cs
@@ -1,37 +1,0 @@
-﻿using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web.Http;
-
-namespace teams_bot2
-{
-	public static class WebApiConfig
-	{
-		public static void Register(HttpConfiguration config)
-		{
-			// Json settings
-			config.Formatters.JsonFormatter.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
-			config.Formatters.JsonFormatter.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
-			config.Formatters.JsonFormatter.SerializerSettings.Formatting = Formatting.Indented;
-			JsonConvert.DefaultSettings = () => new JsonSerializerSettings()
-			{
-				ContractResolver = new CamelCasePropertyNamesContractResolver(),
-				Formatting = Newtonsoft.Json.Formatting.Indented,
-				NullValueHandling = NullValueHandling.Ignore,
-			};
-
-			// Web API configuration and services
-
-			// Web API routes
-			config.MapHttpAttributeRoutes();
-
-			config.Routes.MapHttpRoute(
-					name: "DefaultApi",
-					routeTemplate: "api/{controller}/{id}",
-					defaults: new { id = RouteParameter.Optional }
-			);
-		}
-	}
-}
