--- Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/01 - teams-bot2/teams-bot2/Global.asax.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Http;
-using System.Web.Routing;
-
-namespace teams_bot2
-{
-	public class WebApiApplication : System.Web.HttpApplication
-	{
-		protected void Application_Start()
-		{
-			GlobalConfiguration.Configure(WebApiConfig.Register);
-		}
-	}
-}
