+++ Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/01-advanced-bot/solution/teams-m5-bot/teams_m5_bot.bot
@@ -1,0 +1,14 @@
+{
+  "name": "teams_m5_bot",
+  "secretKey": "",
+  "services": [
+    {
+      "appId": "",
+      "id": "http://localhost:3978/api/messages",
+      "type": "endpoint",
+      "appPassword": "",
+      "endpoint": "http://localhost:3978/api/messages",
+      "name": "teams_m5_bot"
+    }
+  ]
+}
